requirements/test.txt
pytest
pytest-django
pytest-mock
pytest-asyncio
psycopg[binary]
